fix(player): fix bug isTrigger colliders
Player would be able to go on top of isTrigger colliders. This was
because of the raycast-based wheels. In order to prevent this bug in
the future, set the isTrigger collider's layer to `Ignore Raycast`.

Also, along with this bug, the Player gets its Rigidbody and Box
Collider in the same place.